+++ packages/interactivity-router/src/index.ts
@@ -6,7 +6,7 @@ import { store, privateApis, getConfig } from '@wordpress/interactivity';
 /**
  * Internal dependencies
  */
-import { generateCSSStyleSheets } from './assets/styles';
+import { fetchHeadAssets, updateHead, headElements } from './head';
 
 const {
 	directivePrefix,
@@ -37,18 +37,16 @@ interface PrefetchOptions {
 
 interface VdomParams {
 	vdom?: typeof initialVdom;
-	baseUrl?: string;
 }
 
 interface Page {
 	regions: Record< string, any >;
-	styles: Promise< CSSStyleSheet >[];
-	scriptModules: string[];
+	head: HTMLHeadElement[];
 	title: string;
 	initialData: any;
 }
 
-type RegionsToVdom = ( dom: Document, params?: VdomParams ) => Page;
+type RegionsToVdom = ( dom: Document, params?: VdomParams ) => Promise< Page >;
 
 // Check if the navigation mode is full page or region based.
 const navigationMode: 'regionBased' | 'fullPage' =
@@ -75,7 +73,7 @@ const fetchPage = async ( url: string, { html }: { html: string } ) => {
 			html = await res.text();
 		}
 		const dom = new window.DOMParser().parseFromString( html, 'text/html' );
-		return regionsToVdom( dom, { baseUrl: url } );
+		return regionsToVdom( dom );
 	} catch ( e ) {
 		return false;
 	}
@@ -83,17 +81,12 @@ const fetchPage = async ( url: string, { html }: { html: string } ) => {
 
 // Return an object with VDOM trees of those HTML regions marked with a
 // `router-region` directive.
-const regionsToVdom: RegionsToVdom = ( dom, { vdom, baseUrl } = {} ) => {
+const regionsToVdom: RegionsToVdom = async ( dom, { vdom } = {} ) => {
 	const regions = { body: undefined };
-	const styles = generateCSSStyleSheets( dom, baseUrl );
-	const scriptModules = [
-		...dom.querySelectorAll< HTMLScriptElement >(
-			'script[type=module][src]'
-		),
-	].map( ( s ) => s.src );
-
+	let head: HTMLElement[];
 	if ( globalThis.IS_GUTENBERG_PLUGIN ) {
 		if ( navigationMode === 'fullPage' ) {
+			head = await fetchHeadAssets( dom );
 			regions.body = vdom
 				? vdom.get( document.body )
 				: toVdom( dom.body );
@@ -110,28 +103,15 @@ const regionsToVdom: RegionsToVdom = ( dom, { vdom, baseUrl } = {} ) => {
 	}
 	const title = dom.querySelector( 'title' )?.innerText;
 	const initialData = parseServerData( dom );
-	return { regions, styles, scriptModules, title, initialData };
+	return { regions, head, title, initialData };
 };
 
 // Render all interactive regions contained in the given page.
 const renderRegions = async ( page: Page ) => {
-	// Wait for styles and modules to be ready.
-	await Promise.all( [
-		...page.styles,
-		...page.scriptModules.map(
-			( src ) => import( /* webpackIgnore: true */ src )
-		),
-	] );
-	// Replace style sheets.
-	const sheets = await Promise.all( page.styles );
-	window.document
-		.querySelectorAll( 'style,link[rel=stylesheet]' )
-		.forEach( ( element ) => element.remove() );
-	window.document.adoptedStyleSheets = sheets;
-
 	if ( globalThis.IS_GUTENBERG_PLUGIN ) {
 		if ( navigationMode === 'fullPage' ) {
-			// Update HTML.
+			// Once this code is tested and more mature, the head should be updated for region based navigation as well.
+			await updateHead( page.head );
 			const fragment = getRegionRootFragment( document.body );
 			batch( () => {
 				populateServerData( page.initialData );
@@ -189,14 +169,23 @@ window.addEventListener( 'popstate', async () => {
 // Initialize the router and cache the initial page using the initial vDOM.
 // Once this code is tested and more mature, the head should be updated for
 // region based navigation as well.
+if ( globalThis.IS_GUTENBERG_PLUGIN ) {
+	if ( navigationMode === 'fullPage' ) {
+		// Cache the scripts. Has to be called before fetching the assets.
+		[].map.call(
+			document.querySelectorAll( 'script[type="module"][src]' ),
+			( script ) => {
+				headElements.set( script.getAttribute( 'src' ), {
+					tag: script,
+				} );
+			}
+		);
+		await fetchHeadAssets( document );
+	}
+}
 pages.set(
 	getPagePath( window.location.href ),
-	Promise.resolve(
-		regionsToVdom( document, {
-			vdom: initialVdom,
-			baseUrl: window.location.href,
-		} )
-	)
+	Promise.resolve( regionsToVdom( document, { vdom: initialVdom } ) )
 );
 
 // Check if the link is valid for client-side navigation.