+++ test/e2e/specs/interactivity/fixtures/interactivity-utils.ts
@@ -6,30 +6,6 @@ import type { RequestUtils } from '@wordpress/e2e-test-utils-playwright';
 type AddPostWithBlockOptions = {
 	alias?: string;
 	attributes?: Record< string, any >;
-	innerBlocks?: Block[];
-};
-
-type Block = [
-	type: string,
-	attributes?: Record< string, any >,
-	innerBlocks?: Block[],
-];
-
-const generateBlockMarkup = ( [
-	type,
-	attributes,
-	innerBlocks,
-]: Block ): string => {
-	const typeAndAttributes = attributes
-		? `${ type } ${ JSON.stringify( attributes ) }`
-		: type;
-
-	if ( ! innerBlocks ) {
-		return `<!-- wp:${ typeAndAttributes } /-->`;
-	}
-	return `<!-- wp:${ typeAndAttributes } -->${ innerBlocks
-		.map( generateBlockMarkup )
-		.join( '' ) }<!--/ wp:${ type } -->`;
 };
 
 export default class InteractivityUtils {
@@ -64,7 +40,7 @@ export default class InteractivityUtils {
 
 	async addPostWithBlock(
 		name: string,
-		{ attributes, alias, innerBlocks }: AddPostWithBlockOptions = {}
+		{ attributes, alias }: AddPostWithBlockOptions = {}
 	) {
 		const block = attributes
 			? `${ name } ${ JSON.stringify( attributes ) }`
@@ -74,14 +50,8 @@ export default class InteractivityUtils {
 			alias = block;
 		}
 
-		const content = generateBlockMarkup( [
-			name,
-			attributes,
-			innerBlocks,
-		] );
-
 		const payload = {
-			content,
+			content: `<!-- wp:${ block } /-->`,
 			status: 'publish' as 'publish',
 			date_gmt: '2023-01-01T00:00:00',
 			title: alias,